[StreamingPlayer - Edit]
- 기존 코드 파일별로 관리
