+++ ChStreamingPlayer/Extension/UIViewControllerExtension.swift
@@ -1,0 +1,29 @@
+//
+//  UIViewController.swift
+//  ChStreamingPlayer
+//
+//  Created by Chris Kim on 9/28/22.
+//
+
+import Foundation
+import UIKit
+
+
+fileprivate let timeRemainingFormatter: DateComponentsFormatter = {
+    let formatter = DateComponentsFormatter()
+    formatter.zeroFormattingBehavior = .pad
+    formatter.allowedUnits = [.minute, .second]
+    return formatter
+}()
+
+
+extension UIViewController {
+    
+    // MARK: - Convert Time to String
+    
+    func createTimeString(time: Float) -> String {
+        let components = NSDateComponents()
+        components.second = Int(max(0.0, time))
+        return timeRemainingFormatter.string(from: components as DateComponents)!
+    }
+}
