+++ ChStreamingPlayer/Main.storyboard
@@ -176,7 +176,7 @@
                                         <state key="normal" title="Button"/>
                                         <buttonConfiguration key="configuration" style="plain" image="play.fill" catalog="system"/>
                                         <connections>
-                                            <action selector="togglePlay:" destination="3qe-GV-xjU" eventType="touchUpInside" id="LHk-GB-iHQ"/>
+                                            <action selector="togglePlay:" destination="3qe-GV-xjU" eventType="touchUpInside" id="NZ7-MR-2Qa"/>
                                         </connections>
                                     </button>
                                     <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="xsn-Xn-auf">
